main.c rewrite initalization
Blink RED version number 2.1.1
As each component is initialized, if it fails set a bit in init_status (as was originally intended).
If any failures occurred blink RED 13 times so Quality Assurance knows there a problem.
Send out a URL packet with hex bitmap 
Send out a data packet (format x'13' , similar to 03) with "I did bad" ,  the status bits and the battery voltage.

The remaining functionally of main is unchanged. (with the exception of the WDT which causes a fault if enabled!)
Tested with 3 different batteries (2 old, 1 new) and successfully observed correct functionality.


This is the first of several updates.